more cleanups
* add formatting and a few links to readme
* return proper API error message if no auth_token is given
* validate for presence of an access_token
* autogenerate api_token for specs
* add guard clauses where useful
* replace sql fragments with ruby syntax to allow database independence
* use File.join instead manual and error prone path concatenation
* add some table presence checks to allow painless migration up- and downgrading

+++ db/migrate/20130812172949_create_user_aliases.rb
@@ -3,11 +3,11 @@ class CreateUserAliases < ActiveRecord::Migration
     create_table :virtual_user_aliases do |t|
       t.integer :virtual_user_id
       t.string :name
-      t.timestamps
+      t.timestamps null: false
     end
-    
-    drop_view  :aliases_view
-    drop_table :virtual_aliases
+
+    drop_view  :aliases_view if table_exists? :aliases_view
+    drop_table :virtual_aliases if table_exists? :virtual_aliases
   end
 
   def self.down
